+++ dashboard/src/Enum/ProvisioningAction.php
@@ -6,4 +6,5 @@ enum ProvisioningAction: string
 {
     case Create = 'create';
     case Delete = 'delete';
+    case Reset = 'reset';
 }